QUIC: moved ngx_quic_address_hash().
